Move `useEffect` for `myRepos` to `Repos FC`
Went fine

Addresses #49 (Breakout `useEffect` calls into individual components)
Project: 'Breakdown `Cockpit` into subcomponents'
